local-challenger/requirements.txt
requests~=2.32.4
umsgpack~=0.1.0
numpy~=2.1.3
scikit-learn~=1.7.0
pillow~=11.2.1
kafka-python~=2.2.13